ocr-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
easyocr==1.7.1
Pillow==10.1.0
numpy==1.26.2